Improve macros
Emote shortcuts have been formalized as a "built-in macro".
The prefix is changed from `.` to `!`, and the configuration is updated.
Macros can be used on any chat stream now.
A hook has been added for custom macros.
Emotes now use an object, with a callback for handling and whether they should be enabled.